requirements.txt
python-dotenv
aiogram # или python-telegram-bot
SQLAlchemy # или psycopg2, sqlite3
APScheduler
matplotlib
pandas